Move benchmark comparison table to compile directory
