+++ site/frontend/src/pages/compare/compile/test-cases-table.vue
@@ -3,7 +3,7 @@ import {TestCaseComparison} from "../data";
 import Tooltip from "../tooltip.vue";
 import {ArtifactDescription} from "../types";
 import {percentClass} from "../shared";
-import {CompileTestCase} from "../compile/common";
+import {CompileTestCase} from "./common";
 
 const props = defineProps<{
   id: string;
